Laboratorio2/2048log_version 1.1/requirements.txt
numpy
pyparsing
flask
pyswip==0.2.9
jinja2
